create a template directory and an esi_leap_query_events endpoint
