+++ setup.cfg
@@ -41,6 +41,7 @@ console_scripts =
     esi-leap-api = esi_leap.cmd.api:main
     esi-leap-dbsync = esi_leap.cmd.dbsync:main
     esi-leap-manager = esi_leap.cmd.manager:main
+    esi-leap-email-notification = esi_leap.send_email_notification:main
 
 wsgi_scripts =
     esi-leap-api-wsgi = esi_leap.api.wsgi:initialize_wsgi_app
